Add files via upload
this project is a simple text classification project using Naive Bayes for sentiment analysis documents to positive and negative. first the train was on movies reviews data and i tested it in the end on new documents and on book review review to see how much accurate it is and see the likelihood.